Add the entirety of multibar
Still missing good docs, examples and move_cursor_up, but fully
funxional

The mpsc implementation should work for this usecase, even if it locks a
little bit too much (but it'd require a much more complicated API I
really cba to make). Also it's much more a "queue but with yelling"
than a real mpsc but it works so does that matter?